Allow for fetch credentials to be set
In cases where webdav-client is used in the browser natively and Cookies are
used for authentication instead, let users set the `credentials` option when
making a request. If not set, will be made "undefined" and follow default
fetch behavior.

It's recommended NOT to use other authentication means (Basic header, etc.)
along with this, as the results may be undesirable/confusing/undefined/bad.